planqtn_api/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic[standard]
pydantic-settings
python-dotenv
flower
celery
requests
redis
